requirements.txt
google-api-python-client==2.0.2
google-cloud-bigquery==2.11.0
flask==1.1.2
slackclient==2.9.3
pydash==4.9.3
requests==2.25.1
functions-framework==2.1.2
sqlalchemy==1.3.23
psycopg2-binary==2.8.6
google-cloud-storage==1.36.1
nltk==3.5
keras==2.4.3
tensorflow==2.4.1
h5py==2.10.0
spacy==3.0.5
numpy==1.19.5
keras==2.4.3
tensorflow==2.4.1